Rename new theme and theme imports
