+++ client/web/src/Layout.tsx
@@ -8,7 +8,7 @@ import { TabbedPanelContent } from '@sourcegraph/branded/src/components/panel/Ta
 import { FetchFileParameters } from '@sourcegraph/shared/src/backend/file'
 import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { useKeyboardShortcut } from '@sourcegraph/shared/src/keyboardShortcuts/useKeyboardShortcut'
-import { useTheme, Theme } from '@sourcegraph/shared/src/new-theme'
+import { useTheme, Theme } from '@sourcegraph/shared/src/theme'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
 import { Shortcut } from '@sourcegraph/shared/src/react-shortcuts'
 import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
