+++ client/web/src/LegacyLayout.tsx
@@ -9,7 +9,7 @@ import { isMacPlatform } from '@sourcegraph/common'
 import { FetchFileParameters } from '@sourcegraph/shared/src/backend/file'
 import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { useKeyboardShortcut } from '@sourcegraph/shared/src/keyboardShortcuts/useKeyboardShortcut'
-import { useTheme, Theme } from '@sourcegraph/shared/src/new-theme'
+import { useTheme, Theme } from '@sourcegraph/shared/src/theme'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
 import { Shortcut } from '@sourcegraph/shared/src/react-shortcuts'
 import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
