+++ client/web/src/components/externalServices/ExternalServiceForm.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback } from 'react'
 
 import { ErrorLike } from '@sourcegraph/common'
-import { useIsLightTheme } from '@sourcegraph/shared/src/new-theme'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
+import { useIsLightTheme } from '@sourcegraph/shared/src/theme'
 import {
     Button,
     LoadingSpinner,
