+++ client/web/src/enterprise/codeintel/configuration/components/InferenceScriptEditor.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent, useCallback, useMemo, useState } from 'react'
 
-import { useIsLightTheme } from '@sourcegraph/shared/src/new-theme'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
+import { useIsLightTheme } from '@sourcegraph/shared/src/theme'
 import { LoadingSpinner, PageHeader, screenReaderAnnounce, ErrorAlert } from '@sourcegraph/wildcard'
 
 import { AuthenticatedUser } from '../../../../auth'
